Add features from the "further updates" list

+++ backend/src/main/java/com/article/backend/repository/ArticleTagRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface ArticleTagRepository extends JpaRepository<ArticleTag, Integer> {
+public interface ArticleTagRepository extends JpaRepository<ArticleTag, Long> {
     Optional<ArticleTag> findFirstByTagAndArticle_Id(String tag, Long articleId);
 }
